Physics stuff Float -> Double
Former-commit-id: 54ccee75769691aa6aaa2416767bd15270f63347
Former-commit-id: 61e892434f48adf11e3ca49b60928a930fd5cde9

+++ src/net/torvald/point/Point2d.kt
@@ -3,11 +3,11 @@ package net.torvald.point
 /**
  * Created by minjaesong on 16-01-15.
  */
-class Point2f(x: Float, y: Float) {
+class Point2d(x: Double, y: Double) {
 
-    var x: Float = 0.toFloat()
+    var x: Double = 0.toDouble()
         private set
-    var y: Float = 0.toFloat()
+    var y: Double = 0.toDouble()
         private set
 
     init {
@@ -15,7 +15,7 @@ class Point2f(x: Float, y: Float) {
         this.y = y
     }
 
-    operator fun set(x: Float, y: Float) {
+    operator fun set(x: Double, y: Double) {
         this.x = x
         this.y = y
     }
